Project plugins structure refactored
